Refactored react apps components to be cleaner and more segregated
