+++ frontend/src/components/PasswordContainer/PasswordContainer.js
@@ -2,9 +2,8 @@ import { useState, useEffect } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import CircularProgress from '@material-ui/core/CircularProgress';
 import 'react-toastify/dist/ReactToastify.css'
-import { deletePassword, displayPasswords, editPasssword, retrieveUsersPasswords } from '../axiosCalls.js'
+import { retrieveUsersPasswords } from '../axiosCalls.js'
 import Grid from '@material-ui/core/Grid'
-import Typography from '@material-ui/core/Typography'
 import '@fontsource/roboto';
 
 // styling  component
