+++ frontend/src/components/PasswordVault/PasswordVault.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import PasswordComposer from '../PasswordComposer/PasswordComposer';
 import PasswordContainer from '../PasswordContainer/PasswordContainer';
 import 'react-toastify/dist/ReactToastify.css';
